+++ assets/scripts/patentes/data/categories.js
@@ -1,0 +1,11 @@
+/**
+ * Dictionary of diplomatic license plate category codes
+ * Maps one-letter category codes to their descriptions
+ */
+export const categories = {
+    "D": "CUERPO DIPLOMATICO",
+    "I": "ORGANISMO INTERNACIONAL",
+    "C": "CUERPO CONSULAR",
+    "A": "PERSONAL ADMINISTRATIVO",
+    "M": "MISION ESPECIAL"
+};
