+++ assets/scripts/patentes/data/countries.js
@@ -1,0 +1,130 @@
+/**
+ * Dictionary of diplomatic license plate country codes
+ * Maps two-letter country codes to their full country names
+ */
+export const countries = {
+    "AA": "SANTA SEDE",
+    "AC": "REPÚBLICA FEDERAL DE ALEMANIA",
+    "AD": "REPÚBLICA DE ANGOLA",
+    "AE": "REINO DE ARABIA SAUDITA",
+    "AF": "REPÚBLICA ARGELINA DEMOCRÁTICA Y POPULAR",
+    "AG": "REPÚBLICA DE ARMENIA",
+    "AH": "MANCOMUNIDAD DE AUSTRALIA",
+    "AI": "REPÚBLICA DE AUSTRIA",
+    "AJ": "REPÚBLICA DE BELARÚS",
+    "AK": "REINO DE BÉLGICA",
+    "AL": "ESTADO PLURINACIONAL DE BOLIVIA",
+    "AM": "REPÚBLICA FEDERATIVA DEL BRASIL",
+    "AN": "REPÚBLICA DE BULGARIA",
+    "AO": "CANADÁ",
+    "AP": "REPÚBLICA CHECA",
+    "AQ": "REPÚBLICA DE CHILE",
+    "AR": "REPÚBLICA POPULAR CHINA",
+    "AS": "REPÚBLICA DE COLOMBIA",
+    "AT": "REPÚBLICA DEMOCRÁTICA DEL CONGO",
+    "AU": "REPÚBLICA DE COREA",
+    "AV": "REPÚBLICA DE COSTA RICA",
+    "AW": "REPÚBLICA DE CROACIA",
+    "AX": "REPÚBLICA DE CUBA",
+    "AY": "REINO DE DINAMARCA",
+    "AZ": "REPÚBLICA DOMINICANA",
+    "BA": "REPÚBLICA DEL ECUADOR",
+    "BB": "REPÚBLICA ÁRABE DE EGIPTO",
+    "BC": "REPÚBLICA DE EL SALVADOR",
+    "BD": "EMIRATOS ÁRABES UNIDOS",
+    "BE": "REPÚBLICA ESLOVACA",
+    "BF": "REPÚBLICA DE ESLOVENIA",
+    "BG": "REINO DE ESPAÑA",
+    "BH": "ESTADOS UNIDOS DE AMÉRICA",
+    "BI": "REPÚBLICA DE FILIPINAS",
+    "BJ": "REPÚBLICA DE FINLANDIA",
+    "BK": "REPÚBLICA FRANCESA",
+    "BL": "REPÚBLICA HELÉNICA",
+    "BM": "REPÚBLICA DE GUATEMALA",
+    "BN": "REPÚBLICA DE HAITÍ",
+    "BO": "REPÚBLICA DE HONDURAS",
+    "BP": "HUNGRÍA",
+    "BQ": "REPÚBLICA DE LA INDIA",
+    "BR": "REPÚBLICA DE INDONESIA",
+    "BS": "REPÚBLICA ISLÁMICA DE IRÁN",
+    "BT": "IRLANDA",
+    "BU": "ESTADO DE ISRAEL",
+    "BV": "REPÚBLICA ITALIANA",
+    "BW": "JAPÓN",
+    "BX": "ESTADO DE KUWAIT",
+    "BY": "REPÚBLICA LIBANESA",
+    "BZ": "ESTADO DE LIBIA",
+    "CB": "MALASIA",
+    "CC": "SOBERANA ORDEN MILITAR DE MALTA",
+    "CD": "REINO DE MARRUECOS",
+    "CE": "ESTADOS UNIDOS MEXICANOS",
+    "CF": "REPÚBLICA DE NICARAGUA",
+    "CG": "REPÚBLICA FEDERAL DE NIGERIA",
+    "CH": "REINO DE NORUEGA",
+    "CI": "NUEVA ZELANDIA",
+    "CJ": "REINO DE LOS PAÍSES BAJOS",
+    "CK": "REPÚBLICA ISLÁMICA DE PAKISTÁN",
+    "CL": "REPÚBLICA DE PANAMÁ",
+    "CM": "REPÚBLICA DEL PARAGUAY",
+    "CN": "REPÚBLICA DEL PERÚ",
+    "CO": "REPÚBLICA DE POLONIA",
+    "CP": "REPÚBLICA PORTUGUESA",
+    "CQ": "REINO UNIDO DE GRAN BRETAÑA E IRLANDA DEL NORTE",
+    "CR": "RUMANIA",
+    "CS": "FEDERACIÓN DE RUSIA",
+    "CU": "REPÚBLICA DE SERBIA",
+    "CV": "REPÚBLICA ÁRABE SIRIA",
+    "CW": "REPÚBLICA DE SUDÁFRICA",
+    "CX": "REINO DE SUECIA",
+    "CY": "CONFEDERACIÓN SUIZA",
+    "CZ": "REINO DE TAILANDIA",
+    "DA": "REPÚBLICA TUNECINA",
+    "DB": "REPÚBLICA DE TURQUÍA",
+    "DC": "UCRANIA",
+    "DD": "REPÚBLICA ORIENTAL DEL URUGUAY",
+    "DE": "REPÚBLICA BOLIVARIANA DE VENEZUELA",
+    "DF": "REPÚBLICA SOCIALISTA DE VIETNAM",
+    "DG": "UNIÓN EUROPEA",
+    "DH": "LIGA DE LOS ESTADOS ÁRABES",
+    "DI": "ESTADO DE PALESTINA",
+    "DJ": "REPÚBLICA DE AZERBAIYÁN",
+    "DK": "GEORGIA",
+    "DL": "ESTADO DE QATAR",
+    "DM": "MONTENEGRO",
+    "RA": "ALTO COMISIONADO DE LAS NACIONES UNIDAS PARA LOS REFUGIADOS",
+    "RB": "BANCO INTERAMERICANO DE DESARROLLO",
+    "RC": "BANCO INTERNACIONAL DE RECONSTRUCCIÓN Y DESARROLLO",
+    "RD": "CENTRO DE INFORMACIÓN DE LAS NACIONES UNIDAS PARA ARGENTINA Y URUGUAY",
+    "RE": "COMISIÓN ADMINISTRADORA DEL RIO DE LA PLATA",
+    "RF": "COMISIÓN ADMINISTRADORA DEL RÍO URUGUAY",
+    "RG": "COMISIÓN MIXTA ARGENTINO PARAGUAYA DEL RÍO PARANÁ",
+    "RH": "COMITÉ INTERNACIONAL DE LA CRUZ ROJA",
+    "RI": "CORPORACIÓN ANDINA DE FOMENTO",
+    "RJ": "CORPORACIÓN FINANCIERA INTERNACIONAL",
+    "RK": "ENTIDAD BINACIONAL YACIRETÁ",
+    "RL": "FONDO DE LAS NACIONES UNIDAS PARA LA INFANCIA",
+    "RM": "FONDO MONETARIO INTERNACIONAL",
+    "RN": "INSTITUTO INTERAMERICANO DE COOPERACIÓN PARA LA AGRICULTURA",
+    "RO": "ORGANIZACIÓN DE ESTADOS IBEROAMERICANOS PARA LA EDUCACIÓN, LA CIENCIA Y LA CULTURA",
+    "RP": "ORGANIZACIÓN DE LAS NACIONES UNIDAS PARA LA AGRICULTURA Y LA ALIMENTACIÓN",
+    "RQ": "ORGANIZACIÓN IBEROAMERICANO DE SEGURIDAD SOCIAL",
+    "RR": "ORGANIZACIÓN INTERNACIONAL DE POLICIA CRIMINAL",
+    "RS": "ORGANIZACIÓN INTERNACIONAL DEL TRABAJO",
+    "RT": "ORGANIZACION INTERNACIONAL PARA LAS MIGRACIONES",
+    "RU": "ORGANIZACIÓN PANAMERICANA DE LA SALUD - OPS/OMS",
+    "RV": "PROGRAMA COMÚN DE LAS NACIONES UNIDAS SOBRE VIH/SIDA",
+    "RW": "SISTEMA DE LAS NACIONES UNIDAS",
+    "RX": "REPRES. SECRET. GRAL. OEA EN ARGENTINA",
+    "RY": "SECRETARÍA DEL TRATADO ANTÁRTICO",
+    "RZ": "FEDERACIÓN INTERNACIONAL DE SOCIEDADES DE LA CRUZ ROJA",
+    "SA": "REPRESENTACIÓN REGIONAL PARA LAS AMÉRICAS DE LA ORGANIZACIÓN MUNDIAL DE SANIDAD ANIMAL",
+    "SB": "OFICINA DE LAS NACIONES UNIDAS PARA LOS PROYECTOS",
+    "SC": "CENTRO DE ESTUDIOS ESTRATÉGICOS DE DEFENSA DEL CONSEJO DE DEFENSA SURAMERICANO DEL UNASUR - CEED",
+    "SD": "COMISIÓN TRINACIONAL PARA EL DESARROLLO DE LA CUENCA DEL RÍO PILCOMAYO",
+    "SE": "COMISIÓN ECONÓMICA PARA AMÉRICA LATINA Y EL CARIBE - CEPAL",
+    "XA": "AGENCIA DE COOPERACIÓN INTERNACIONAL DEL JAPÓN (JICA)",
+    "XB": "FUNDACIÓN KONRAD ADENAUER",
+    "XC": "FUNDACIÓN HANNS SEIDEL",
+    "XD": "FUNDACIÓN FRIEDRICH NAUMANN",
+    "XE": "FUNDACIÓN FRIEDRICH EBERT"
+};
